Выполняет 5 задание, разбивает на модули
